refactor: split cashflow planning ingestion dtos
